feat(library): 书/漫画/视频三域统一「库页三视图」导航 + 漫画本地扫描根
用户诉求：漫画 tab 抄 Mangatan 的 [书架|浏览|漫画源] 分段导航，且「管理来源」
不该藏在页头按钮后面——它是一等视图；三个媒体域都要有同一套结构。

导航壳（MediaLibraryShell）
- 各域只声明自己真有的视图：漫画=书架/浏览/来源，书=书架/来源（小说在线源
  未落地，不放空壳 tab），视频=媒体库/来源（动画走 torrent，无在线浏览源）。
  只有一个视图时连导航条都不显示，调用侧无需写条件分支。
- 壳不外包页头：三个库页各自带标题与动作（导入/合集/统计），外层再加页头就是
  双标题。分段条下沉到各视图页头的 bottom 槽——标题动作原样，位置全域一致。
- 分段条必须包 HibikiAdjustableSegmented：裸 SegmentedButton 只是一堆原生按钮，
  只遍历已注册 target 的方向焦点控制器会整个跳过，手柄/键盘用户切不了视图。
- 视图惰性构建 + Offstage/TickerMode 保活（与 home_page 顶层 tab 同源）：没访问
  过的不构建（在线目录不会因挂载就发网络请求），访问过的切走保留滚动位置。

来源管理提升为一等视图
- MediaSourcesDialog 的实现体提取为 MediaSourcesView，对话框退为薄壳，新增
  MediaSourcesPage 消费同一份行为——两种语境的差别只剩 chrome（添加按钮在页脚
  还是页头）。避免「书架/视频页各写一份」的老路。
- 页头「管理来源」按钮只在页面被独立使用（无导航条）时保留；在壳里已是视图，
  两个入口就是同一件事的两份实现。
- 书架页的对话框入口按 mangaOnly 选 kind（此前漫画书架根本没有来源入口）。

漫画本地扫描根（SourceLibraryKind.manga）
- 枚举加 manga、countMediaBySourceId 加 manga 分支（epub_books 且 format='manga'）。
- 扫描器 planScan 认 .mokuro，穷尽 switch 加 manga case，逐卷走
  MangaImporter.importFromMokuroPath(skipIfExists: true)，重复静默跳过（BUG-443
  范式）。首版只认 .mokuro：CBZ/裸图没有 skipIfExists 身份契约，重扫会重复导入。
- 网络 manga 源按 video 先例拒绝（卷体积大且导入要解包，远端流式扫描无意义）。

在线目录（mokuro.moe）内容体同样提取为 MokuroMoeCatalogView，下载页的对话框
入口与漫画 tab 的「浏览」视图共用；下载仍走 app 级共享队列，切视图不中断。

守卫测试跟着实现移位改扫描目标（断言逐条未放宽）；漫画页接线守卫穿过壳层断言，
并新锚定三视图顺序与来源视图的 mediaKind。

验证：flutter analyze（含 test）零问题；定向 6730+ 测试全绿
（pages 2358 / i18n+media+settings+database 4372 / 扫描器+kind 36）。
