+++ hibiki/test/database/source_library_kind_test.dart
@@ -7,13 +7,15 @@ void main() {
   test('dbValue 逐字节冻结', () {
     expect(SourceLibraryKind.book.dbValue, 'book');
     expect(SourceLibraryKind.video.dbValue, 'video');
-    expect(SourceLibraryKind.values, hasLength(2),
+    expect(SourceLibraryKind.manga.dbValue, 'manga');
+    expect(SourceLibraryKind.values, hasLength(3),
         reason: '新增来源库种类必须同步扫描/统计/对话框的穷尽 switch 与本守卫');
   });
 
   test('tryParse:落库串双向、未知/跨域串拒斥', () {
     expect(SourceLibraryKind.tryParse('book'), SourceLibraryKind.book);
     expect(SourceLibraryKind.tryParse('video'), SourceLibraryKind.video);
+    expect(SourceLibraryKind.tryParse('manga'), SourceLibraryKind.manga);
     expect(SourceLibraryKind.tryParse('epub'), isNull,
         reason: 'MediaKind 的值域串不属于来源库值域(book≠epub 是真实语义差)');
     expect(SourceLibraryKind.tryParse('game'), isNull);
