+++ packages/hibiki_core/lib/src/database/source_library_kind.dart
@@ -1,7 +1,8 @@
 /// 来源库(扫描根,`MediaSources` 表)的媒体种类值域(命名统一 Phase 3.4 批 2)。
 ///
-/// ⚠️ 与 [MediaKind](media_kind.dart) **不是**同一值域:来源库只分
-/// 'book'/'video' 两类(book 源既装 EPUB 也装 SRT 书,不细分),`book`≠`epub`
+/// ⚠️ 与 [MediaKind](media_kind.dart) **不是**同一值域:来源库分
+/// 'book'/'video'/'manga' 三类(book 源既装 EPUB 也装 SRT 书,不细分;manga 源
+/// 扫 `.mokuro` 卷,落 `EpubBooks.format='manga'` 行),`book`≠`epub`
 /// 是真实语义差,不做合并——这正是 `MediaSources.mediaKind` 列「字段名叫
 /// mediaKind 却装不下 MediaKind 值」的历史根源,本枚举把它显式化。
 /// 跨域换算见 media_kind_mappings.dart 的既有模式(本值域暂无跨域映射需求)。
@@ -9,7 +10,8 @@
 /// [dbValue] 即落库串,**逐字节冻结**(守卫:source_library_kind_test)。
 enum SourceLibraryKind {
   book('book'),
-  video('video');
+  video('video'),
+  manga('manga');
 
   const SourceLibraryKind(this.dbValue);
